wip, kind pinning is broken
