+++ Example/Example/View/SwiftUI/SwiftUIListNoSections.swift
@@ -6,17 +6,6 @@ struct SwiftUIListNoSections: View {
     typealias Inserter = ListInserter.Inserter<NoSection<BookItem, PromotionalView>>
     typealias Item = ListInserter.Item<BookItem, PromotionalView>
     
-    struct BookItem: Hashable, ValueKindIdentifiable {
-        
-        enum Kind: Hashable {
-            case fantasy(String)
-            case horror(String, Int)
-        }
-            
-        let id = UUID() // allows duplicate entries of the exact same item
-        let valueKind: Kind
-    }
-    
     class Model: ObservableObject {
         private var rawEntries = [BookItem]()
         @Published var entries = [Item]()
@@ -94,13 +83,12 @@ struct SwiftUIListNoSections: View {
             }
             
             List {
-                ForEach(model.entries, id: \.self) { entry in
-                    
+                ForEach(model.entries, id: \.self) { entry in                    
                     switch entry {
                     case let .inserted(insertedItemInfo):
                         insertedItemInfo.embed
-                    case let .value(value):
-                        switch value.valueKind {
+                    case let .value(bookItem):
+                        switch bookItem.details {
                         case let .fantasy(text):
                             Text("text only. text: \(text)")
                                 .foregroundColor(.green)
@@ -111,18 +99,6 @@ struct SwiftUIListNoSections: View {
                     }
                 }
 //                .onDelete(perform: model.deleteItem)
-            }
-        }
-    }
-}
-
-private extension [SwiftUIListNoSections.BookItem] {
-    static func generateRandom(count: Int = 10) -> [Element] {
-        return (0..<count).map { index in
-            if Bool.random() {
-                return Element(valueKind: .fantasy("Lorem text: \(index)"))
-            } else {
-                return Element(valueKind: .horror("Lorem text:", Int.random(in: 100..<1000)))
             }
         }
     }
