+++ Example/Example/View/SwiftUI/SwiftUIListWithSections.swift
@@ -11,17 +11,6 @@ struct SwiftUIListWithSections: View {
         let items: [BookItem]
     }
     
-    struct BookItem: Hashable, ValueKindIdentifiable {
-        
-        enum Kind: Hashable {
-            case fantasy(String)
-            case horror(String, Int)
-        }
-            
-        let id = UUID() // allows duplicate entries of the exact same item
-        let valueKind: Kind
-    }
-    
     class Model: ObservableObject {
         @Published var sections = [ListSection]()
         
@@ -103,30 +92,18 @@ struct SwiftUIListWithSections: View {
                             case let .inserted(insertedItemInfo):
                                 insertedItemInfo.embed
                             case let .value(bookItem):
-                                switch bookItem.valueKind {
-                                    case let .fantasy(text):
-                                        Text("text only. text: \(text)")
-                                            .foregroundColor(.green)
-                                    case let .horror(text, num):
-                                        Text("text and num. text:\(text) and num:\(num)")
-                                            .foregroundColor(.blue)
+                                switch bookItem.details {
+                                case let .fantasy(text):
+                                    Text("text only. text: \(text)")
+                                        .foregroundColor(.green)
+                                case let .horror(text, num):
+                                    Text("text and num. text:\(text) and num:\(num)")
+                                        .foregroundColor(.blue)
                                 }
                             }
                         }
                     }
                 }
-            }
-        }
-    }
-}
-
-private extension [SwiftUIListWithSections.BookItem] {
-    static func generateRandom(count: Int = 10) -> [Element] {
-        return (0..<count).map { index in
-            if Bool.random() {
-                return Element(valueKind: .fantasy("Fantasy book \(Int.random(in: 1..<1_000))"))
-            } else {
-                return Element(valueKind: .horror("Horror book", Int.random(in: 100..<1_000)))
             }
         }
     }
